move ProfileCard component
